main/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
django-bootstrap3==21.2
django-environ==0.8.1
django-modeltranslation==0.18.2
django-php==0.1
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
psycopg2==2.9.3
pyTelegramBotAPI==4.4.0
requests==2.27.1
six==1.16.0
solid-i18n==1.4.2
sqlparse==0.4.2
telepot==12.7
tzdata==2022.1
urllib3==1.26.9
yarl==1.7.2


